Update 20_data.Rmd
Not a big deal, but moving the filter operation to the end allow you to keep the macro variables for January in 1960.  

Thanks for putting this together, its a fantastic resource!  I started by just skimming through the first six or seven chapters, and now I'm going back through and playing with the code.  I've already learned a lot and it's been a good refresher on a bunch of other things.  I'm a lecturer in the finance department at Auburn University, and I definitely plan to show this to our incoming PhD students.  

I'm really glad this exists and if you ever need any help with the project or want someone to review anything or need contributions in any other way, feel free to reach out.  My name is Joseph Stover and my email is jstover@auburn.edu.  I'd be happy to help!